updating styling for the card components
