+++ content/docs/index.mdx
@@ -4,16 +4,41 @@ seo:
   title: Example Title
   description: Example Description
 title: Tina Docs
-last_edited: '2025-04-08T07:37:26.816Z'
+last_edited: '2025-04-28T07:02:12.149Z'
 next: content/docs/core-concepts/overview.mdx
 previous: ''
 ---
 
 ## Introduction
 
+<cardGrid
+  cards={[
+    {
+      title: "Card Title",
+      description: "Card Description",
+      link: "https://www.google.com",
+      linkText: "Search now"
+    },
+    {
+      title: "Card Title",
+      description: "Card Description",
+      link: "https://www.google.com",
+      linkText: "Search now"
+    },
+    { title: "Test", description: "some text" },
+    { title: "Test", description: "some more text" },
+    {
+      title: "This is a really long title.",
+      description:
+        "This is also a long bit of text just to check everything moves as expected."
+    },
+    { title: "Grid business.", description: "Grid business..." }
+  ]}
+/>
+
 Tina is an open-source, Git-backed headless content management system (CMS) that empowers both developers and content creators to collaborate seamlessly on a single platform.
 
- Tina enables developers to craft a custom visual editing experience perfectly tailored to their website or application while supporting a broad range of content types such as Markdown, MDX, and JSON.
+Tina enables developers to craft a custom visual editing experience perfectly tailored to their website or application while supporting a broad range of content types such as Markdown, MDX, and JSON.
 
 <youtube embedSrc="https://www.youtube.com/embed/zRkeKSZjlyw" />
