+++ src/app/error-wrapper.tsx
@@ -33,6 +33,7 @@ const ErrorWrapper = ({
                   <Link
                     href={link.linkUrl}
                     passHref
+                    key={link.linkUrl}
                     className="text-slate-500 shadow-sm hover:shadow-md outline outline-slate-200 hover:text-slate-700 rounded-md p-2 bg-white/50 hover:bg-white/90"
                   >
                     <div>{link.linkText ?? "External Link 🔗"}</div>
