+++ tina/templates/markdown-embeds/card-grid.template.tsx
@@ -1,12 +1,29 @@
 export const CardGridTemplate = {
   name: "cardGrid",
   label: "Card Grid",
+  ui: {
+    defaultItem: () => {
+      return {
+        cards: [],
+      };
+    },
+  },
   fields: [
     {
       name: "cards",
       label: "Cards",
       type: "object",
       list: true,
+      ui: {
+        defaultItem: () => {
+          return {
+            title: "Card Title",
+            description: "Card Description",
+            link: "https://www.google.com",
+            linkText: "Search now",
+          };
+        },
+      },
       fields: [
         {
           name: "title",
